requirements.txt
Flask==3.0.2
google-genai
python-dotenv==1.0.1
cryptography==42.0.5
PyYAML==6.0.1
